move bytes again
